requirements.txt
google-search-py==1.3.1
mysql-connector-python==8.0.33
psutil==5.9.5
python-dotenv==1.0.0
pyttsx3==2.90
requests==2.31.0
selenium==4.14.0
SpeechRecognition==3.10.0